+++ src/NLoad.Tests/Mocks/TestMock.cs
@@ -1,0 +1,18 @@
+using System.Threading;
+
+namespace NLoad.Tests
+{
+    public class TestMock : ITest
+    {
+        public void Initialize()
+        {
+        }
+
+        public TestResult Execute()
+        {
+            Thread.Sleep(1);
+
+            return new TestResult(true);
+        }
+    }
+}
