+++ src/NLoad.Tests/Mocks/ThreadCounter.cs
