+++ src/NLoad.Tests/NLoad.Tests.csproj
@@ -50,12 +50,15 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="ThreadCounter.cs" />
-    <Compile Include="TestMock.cs" />
+    <Compile Include="Mocks\OneSecondDelayTest.cs" />
+    <Compile Include="Mocks\RandomDelayTest.cs" />
+    <Compile Include="Mocks\ThreadCounter.cs" />
+    <Compile Include="Mocks\TestMock.cs" />
     <Compile Include="LoadTestBuilderTests.cs" />
     <Compile Include="LoadTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TestRunnerTests.cs" />
+    <Compile Include="TestRunResultTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\NLoad\NLoad.csproj">
