+++ src/NLoad/LoadTestException.cs
@@ -1,7 +1,9 @@
 ﻿using System;
+using System.Diagnostics.CodeAnalysis;
 
 namespace NLoad
 {
+    [ExcludeFromCodeCoverage]
     public class LoadTestException : Exception
     {
         public LoadTestException(string message)
